utils/schemas/requirements.txt
setuptools==80.9.0
PyYAML==6.0.3
jsonref==1.1.0

